Always delay hover display
