+++ crates/editor/src/editor.rs
@@ -6147,12 +6147,8 @@ pub fn styled_runs_for_code_label<'a>(
 
 #[cfg(test)]
 mod tests {
-    use crate::{
-        hover_popover::{hover, hover_at, HoverAt, HOVER_DELAY_MILLIS, HOVER_GRACE_MILLIS},
-        test::{
-            assert_text_with_selections, build_editor, select_ranges, EditorLspTestContext,
-            EditorTestContext,
-        },
+    use crate::test::{
+        assert_text_with_selections, build_editor, select_ranges, EditorTestContext,
     };
 
     use super::*;
@@ -6164,7 +6160,7 @@ mod tests {
     use indoc::indoc;
     use language::{FakeLspAdapter, LanguageConfig};
     use lsp::FakeLanguageServer;
-    use project::{FakeFs, HoverBlock};
+    use project::FakeFs;
     use settings::LanguageOverride;
     use std::{cell::RefCell, rc::Rc, time::Instant};
     use text::Point;
@@ -9393,193 +9389,6 @@ mod tests {
             .next()
             .await;
         }
-    }
-
-    #[gpui::test]
-    async fn test_hover_popover(cx: &mut gpui::TestAppContext) {
-        let mut cx = EditorLspTestContext::new_rust(
-            lsp::ServerCapabilities {
-                hover_provider: Some(lsp::HoverProviderCapability::Simple(true)),
-                ..Default::default()
-            },
-            cx,
-        )
-        .await;
-
-        // Basic hover delays and then pops without moving the mouse
-        cx.set_state(indoc! {"
-            fn |test()
-                println!();"});
-        let hover_point = cx.display_point(indoc! {"
-            fn test()
-                print|ln!();"});
-
-        cx.update_editor(|editor, cx| {
-            hover_at(
-                editor,
-                &HoverAt {
-                    point: Some(hover_point),
-                },
-                cx,
-            )
-        });
-        assert!(!cx.editor(|editor, _| editor.hover_state.visible()));
-
-        // After delay, hover should be visible.
-        let symbol_range = cx.lsp_range(indoc! {"
-            fn test()
-                [println!]();"});
-        let mut requests =
-            cx.lsp
-                .handle_request::<lsp::request::HoverRequest, _, _>(move |_, _| async move {
-                    Ok(Some(lsp::Hover {
-                        contents: lsp::HoverContents::Markup(lsp::MarkupContent {
-                            kind: lsp::MarkupKind::Markdown,
-                            value: indoc! {"
-                        # Some basic docs
-                        Some test documentation"}
-                            .to_string(),
-                        }),
-                        range: Some(symbol_range),
-                    }))
-                });
-        cx.foreground()
-            .advance_clock(Duration::from_millis(HOVER_DELAY_MILLIS + 100));
-        requests.next().await;
-
-        cx.editor(|editor, _| {
-            assert!(editor.hover_state.visible());
-            assert_eq!(
-                editor.hover_state.popover.clone().unwrap().contents,
-                vec![
-                    HoverBlock {
-                        text: "Some basic docs".to_string(),
-                        language: None
-                    },
-                    HoverBlock {
-                        text: "Some test documentation".to_string(),
-                        language: None
-                    }
-                ]
-            )
-        });
-
-        // Mouse moved with no hover response dismisses
-        let hover_point = cx.display_point(indoc! {"
-            fn te|st()
-                println!();"});
-        cx.update_editor(|editor, cx| {
-            hover_at(
-                editor,
-                &HoverAt {
-                    point: Some(hover_point),
-                },
-                cx,
-            )
-        });
-        cx.lsp
-            .handle_request::<lsp::request::HoverRequest, _, _>(|_, _| async move { Ok(None) })
-            .next()
-            .await;
-        cx.foreground().run_until_parked();
-        cx.editor(|editor, _| {
-            assert!(!editor.hover_state.visible());
-        });
-        cx.foreground()
-            .advance_clock(Duration::from_millis(HOVER_GRACE_MILLIS + 100));
-
-        // Hover with keyboard has no delay
-        cx.set_state(indoc! {"
-            f|n test()
-                println!();"});
-        cx.update_editor(|editor, cx| hover(editor, &hover_popover::Hover, cx));
-        let symbol_range = cx.lsp_range(indoc! {"
-            [fn] test()
-                println!();"});
-        cx.lsp
-            .handle_request::<lsp::request::HoverRequest, _, _>(move |_, _| async move {
-                Ok(Some(lsp::Hover {
-                    contents: lsp::HoverContents::Markup(lsp::MarkupContent {
-                        kind: lsp::MarkupKind::Markdown,
-                        value: indoc! {"
-                        # Some other basic docs
-                        Some other test documentation"}
-                        .to_string(),
-                    }),
-                    range: Some(symbol_range),
-                }))
-            })
-            .next()
-            .await;
-        cx.foreground().run_until_parked();
-        cx.editor(|editor, _| {
-            assert!(editor.hover_state.visible());
-            assert_eq!(
-                editor.hover_state.popover.clone().unwrap().contents,
-                vec![
-                    HoverBlock {
-                        text: "Some other basic docs".to_string(),
-                        language: None
-                    },
-                    HoverBlock {
-                        text: "Some other test documentation".to_string(),
-                        language: None
-                    }
-                ]
-            )
-        });
-
-        // Open hover popover disables delay
-        let hover_point = cx.display_point(indoc! {"
-            fn test()
-                print|ln!();"});
-        cx.update_editor(|editor, cx| {
-            hover_at(
-                editor,
-                &HoverAt {
-                    point: Some(hover_point),
-                },
-                cx,
-            )
-        });
-
-        let symbol_range = cx.lsp_range(indoc! {"
-            fn test()
-                [println!]();"});
-        cx.lsp
-            .handle_request::<lsp::request::HoverRequest, _, _>(move |_, _| async move {
-                Ok(Some(lsp::Hover {
-                    contents: lsp::HoverContents::Markup(lsp::MarkupContent {
-                        kind: lsp::MarkupKind::Markdown,
-                        value: indoc! {"
-                        # Some third basic docs
-                        Some third test documentation"}
-                        .to_string(),
-                    }),
-                    range: Some(symbol_range),
-                }))
-            })
-            .next()
-            .await;
-        cx.foreground().run_until_parked();
-        // No delay as the popover is already visible
-
-        cx.editor(|editor, _| {
-            assert!(editor.hover_state.visible());
-            assert_eq!(
-                editor.hover_state.popover.clone().unwrap().contents,
-                vec![
-                    HoverBlock {
-                        text: "Some third basic docs".to_string(),
-                        language: None
-                    },
-                    HoverBlock {
-                        text: "Some third test documentation".to_string(),
-                        language: None
-                    }
-                ]
-            )
-        });
     }
 
     #[gpui::test]
